src_enterprise/requirements.txt
taipy-enterpise
scikit-learn
pmdarima
plotly